re sacffolded database and removed trainee now user is the trainee

+++ Application/Models/Grade.cs
@@ -9,7 +9,7 @@ public partial class Grade
 
     public int EnrollmentId { get; set; }
 
-    public double Grade1 { get; set; }
+    public decimal Grade1 { get; set; }
 
     public bool? IsPass { get; set; }
